fix: use int for TTL

+++ arq_dashboard/core/settings.py
@@ -40,5 +40,5 @@ MAX_PER_SECONDS: Optional[int] = config(
     "ARQ_DASHBOARD_MAX_PER_SECONDS", cast=int, default=None
 )
 
-CACHE_TTL: float = config("ARQ_DASHBOARD_CACHE_TTL", cast=float, default=60.0)
+CACHE_TTL: int = config("ARQ_DASHBOARD_CACHE_TTL", cast=int, default=60)
 CACHE_MAX_SIZE: int = config("ARQ_DASHBOARD_CACHE_MAX_SIZE", cast=int, default=32)
